payment_api/requirements.txt
fastapi==0.85.1
orjson==3.8.0
pydantic==1.10.2
uvicorn==0.19.0
gunicorn==20.1.0
SQLAlchemy==1.4.42
alembic==1.7.1
asyncpg==0.24.0
async-stripe==4.2.0
pyjwt==2.5.0
python-dateutil==2.8.2
jinja2==3.1.2
python-dotenv==0.21.0
sentry-sdk==1.11.1